pyproject.toml
[tool.poetry]
name = "llm-pdf-poc"
version = "0.1.0"
description = ""
authors = ["Francisco <vieirafrancisco.dev@gmail.com>"]
readme = "README.md"
packages = [{include = "llm_pdf_poc"}]

[tool.poetry.dependencies]
python = "^3.12"
PyPDF2 = "^3.0.1"
google-generativeai = "^0.7.2"
nltk = "^3.9.1"
environs = "^11.0.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
